moving movies list loader into viewmodel
