+++ app/src/main/java/ru/chaichuk/hwapp/MainActivity.kt
@@ -14,8 +14,7 @@ import ru.chaichuk.hwapp.fragments.OnMoviesListClickListener
 class MainActivity :
     AppCompatActivity(),
     OnMoviesListClickListener,
-    OnMoviesDetailsClickListener,
-    MoviesListLoader {
+    OnMoviesDetailsClickListener {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -36,47 +35,4 @@ class MainActivity :
     override fun onDetailsBack() {
         supportFragmentManager.popBackStack()
     }
-
-    override suspend fun loadMoviesList(): List<Movie> =
-        withContext(Dispatchers.IO) {
-            //delay(3_000)
-            //loadMovies(applicationContext)
-
-            val pictureBasePath = "https://image.tmdb.org/t/p/w780"
-            val movieDbApi = MovieDbApi()
-            val movies : MutableList<Movie> = mutableListOf()
-            val moviesDTO = movieDbApi.getPopularMovies()
-            for (movieDTO in moviesDTO) {
-                val movieDetailsDTO = movieDbApi.getMovieDetails(movieDTO.id.toInt())
-                val movieCreditsDTO = movieDbApi.getMovieCredits(movieDTO.id.toInt())
-                val genres : List<Genre> = movieDetailsDTO.genres.map { Genre(it.id.toInt(), it.name) }
-                val actors : List<Actor> = movieCreditsDTO.cast.mapNotNull {
-                    it.profilePath?.let { picture ->
-                        Actor(
-                            it.id.toInt(),
-                            it.name,
-                            pictureBasePath + picture
-                        )
-                    }
-                }
-
-                val movie = Movie (
-                    movieDTO.id.toInt(),
-                    movieDTO.title,
-                    movieDTO.overview,
-                    pictureBasePath + movieDTO.posterPath,
-                    pictureBasePath + movieDTO.backdropPath,
-                    movieDTO.voteAverage.toFloat(),
-                    movieDTO.voteCount.toInt(),
-                    if(movieDTO.adult) 16 else 13,
-                    movieDetailsDTO.runtime.toInt(),
-                    genres,
-                    actors,
-                    true
-                )
-                movies.add(movie)
-            }
-            return@withContext movies
-
-        }
 }
